Port stencil.FieldType and stencil.IndexAttr out of the experimental directory
